requirements.txt
aiofiles>=24.1.0
aiohappyeyeballs>=2.6.1
aiohttp>=3.12.15
aiosignal>=1.4.0
aiosqlite>=0.21.0
argon2-cffi>=25.1.0
argon2-cffi-bindings>=25.1.0
attrs>=25.3.0
certifi>=2026.1.4
charset-normalizer>=3.4.4
discord.py>=2.6.3
filelock>=3.20.2
frozenlist>=1.7.0
idna>=3.10
multidict>=6.6.4
propcache>=0.3.2
python-dotenv>=1.1.1
python-magic>=0.4.14
requests>=2.32.5
requests-file>=3.0.1
tldextract>=5.3.1
typing_extensions>=4.15.0
urllib3>=2.6.2
yarl>=1.20.1